% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrollytell.R
\name{scrolly_container}
\alias{scrolly_container}
\title{Container for scrollytelling}
\usage{
scrolly_container(outputId, ..., width = "100\%", height = "400px")
}
\description{
Container for scrollytelling
}
