% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrollytags.R
\name{scrolly_graph}
\alias{scrolly_graph}
\title{Sticky graph}
\usage{
scrolly_graph(..., width = "60\%", class = "scrolly-graph")
}
\arguments{
\item{...}{elements that will be part of the sticky graph}

\item{width}{Width of the graph in the `scrolly_container`}

\item{class}{The css class of the container element (do not remove)}
}
\description{
`scrolly_graph` should be used within a `scrolly_container`
This will be the "sticky" part that remains in view until the user scrolls beyond the end of
the container
}
