% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrollytags.R
\name{scrolly_section}
\alias{scrolly_section}
\title{Section element}
\usage{
scrolly_section(id, ..., class = "scrolly-section")
}
\arguments{
\item{id}{ID of scrolly_section.}

\item{...}{elements that are part of the section}

\item{css}{class, do not remove}
}
\description{
section element of the scrolly container. When the section element is selected, the value of the container is equal
to the section id.
}
