% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrollytags.R
\name{scrolly_sections}
\alias{scrolly_sections}
\title{container for \code{scrolly_section}s.}
\usage{
scrolly_sections(..., width = "40\%", class = "scrolly-sections")
}
\description{
Creates a part of a graph that will stay in screen, until the end of the container
}
