% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrollytags.R
\name{scrolly_graph}
\alias{scrolly_graph}
\title{Sticky graph}
\usage{
scrolly_graph(..., width = "60\%", class = "scrolly-graph")
}
\description{
Creates a part of a graph that will stay in screen, until the end of the container
}
