% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrollytags.R
\name{scrolly_sections}
\alias{scrolly_sections}
\title{container for \code{scrolly_section}s.}
\usage{
scrolly_sections(..., width = "40\%", class = "scrolly-sections")
}
\arguments{
\item{...}{section elements}

\item{width}{Width of the non-sticky part}

\item{class}{css class used}
}
\description{
The non-sticky part of a `scrolly_container`.
Contains the text `section`s that will scroll when a user scrolls the `scrolly_container`.
}
